Move 'base' serializers
